+++ CHANGES.rst
@@ -2,6 +2,9 @@ Changelog
 =========
 
 
+- Don't create the root directory (if it doesn't exist) until at least one file has been added.
+
+
 v0.3.0 (2015-06-02)
 -------------------